requirements.txt
tensorflow-gpu==1.15.0
numpy==1.19.2
faiss==1.7.1
scipy==1.3.0
tqdm==4.62.3

